finally make boot_control.common become boot_control._ota_status_control module
